+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file. The
 format is based on [Keep a Changelog], and this project aims to follow
 [Semantic Versioning].
 
+## [0.27.0] - 2026-03-06
+
+### Changed
+
+- `TransferState` extended with `as_load`, `as_store`
+
+- `CollectionState` extended with `Error`
+
 ## [0.26.0] - 2026-03-04
 
 ### Changed